+++ packages/twenty-server/src/engine/core-modules/dns-manager/validator/dns-manager.validate.ts
